+++ DiscCheckEmuAPIHook/Config/ConfigParser.cpp
@@ -172,9 +172,9 @@ namespace dce {
 		{
 			dce::Patch p;
 			if (patch["offset"].IsDefined())
-				p.address = memory_utils::getVAFromOffset(patch["offset"].as<std::uintptr_t>());
+				p.address = memory_utils::getVAFromOffset(patch["offset"].as<std::size_t>());
 			else if (patch["address"].IsDefined())
-				p.address = patch["address"].as<std::uintptr_t>();
+				p.address = reinterpret_cast<uint8_t*>(patch["address"].as<std::size_t>());
 			else
 				throw std::exception("Error: Patches should have an offset or an address");